rename waiting screen to lobby screen
