+++ client-mobile/components/index.tsx
@@ -3,5 +3,5 @@ export { default as GameScreen } from "./GameScreen";
 export { default as LoadingScreen } from "./LoadingScreen";
 export { default as RegisterScreen } from "./RegisterScreen";
 export { default as StartScreen } from "./StartScreen";
-export { default as WaitingScreen } from "./WaitingScreen";
+export { default as LobbyScreen } from "./LobbyScreen";
 export { default as CustomCardScreen } from "./CustomCardScreen";
